software/requirements.txt
flask
flask-restful
paho-mqtt